moved joypad and apu features back to the dmg file for consistency
