+++ includes/memory/mmio/mmio.hpp
@@ -33,6 +33,8 @@ enum class IORegisterMapping : addr_t {
   MMIO_TIMER_TMA = 0xFF06,
   MMIO_TIMER_TAC = 0xFF07,
   MMIO_INT_FLAGS = 0xFF0F,
+  MMIO_AUDIO_BASE = 0xFF10,
+  MMIO_WAVE_RAM_BASE = 0xFF30,
   MMIO_LCD_CONTROL = 0xFF40,
   MMIO_LCD_STATUS = 0xFF41,
   MMIO_LCD_SCY = 0xFF42,
